Fix minor errors detected by CRAN, work around Xvfb bugs
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/rgl/pkg/rgl@1197 79a1b648-8b68-254f-b20b-c0b50f5137bf

+++ R/bgplot3d.R
@@ -11,11 +11,16 @@ bgplot3d <- function(expression) {
   viewport <- par3d("viewport")
   width <- viewport["width"]
   height <- viewport["height"]
-  filename <- tempfile(fileext = ".png")
-  png(filename = filename, width=width, height=height)
-  value <- expression  
-  dev.off()
-  result <- bg3d(texture=filename, col="white")
-  unlink(filename)
+  if (width > 0 && height > 0) {
+    filename <- tempfile(fileext = ".png")
+    png(filename = filename, width=width, height=height)
+    value <- try(expression)  
+    dev.off()
+    result <- bg3d(texture=filename, col="white")
+    unlink(filename)
+  } else {
+    value <- NULL
+    result <- bg3d(col="white")
+  }
   invisible(structure(result, value = value))
 }
